requirement.txt
google-generativeai
PIL
gtts
streamlit